Restore the custom build script for pyarrow
For local development it's much faster to be able do python setup.py develop
So we use the previous thirdparty build script to build pyarrow against
system libarrow that we installed to enable liblance build.

The main change in this PR is changing the arrow ref being checked out. By
default this builds pyarrow 8.0.0.devXXX, which fails the pylance requirement
of pyarrow>=8,<9. Instead we use the 8.0.1 tag in pyarrow