pandas image extension
We have the Image arrow extension type but it breaks to_pandas conversion.
Here we implement a pandas extension dtype for ImageArray with ImageUri's.
It is backed by a string ndarray. Note that this requires copying the strings
which impacts performance a lot. A later optimization could be to directly
back this with arrow array.