Reorganize tutorial notebooks (#265)
* pandas image extension

We have the Image arrow extension type but it breaks to_pandas conversion.
Here we implement a pandas extension dtype for ImageArray with ImageUri's.
It is backed by a string ndarray. Note that this requires copying the strings
which impacts performance a lot. A later optimization could be to directly
back this with arrow array.

* add arrow-pandas roundtrip

* address pr comments

* fix tests on mac

* Reorganize tutorial notebooks

Renamed labels => 3. EDA
Renamed peeking_duck => 4. Model inference (symlinked since it's tweeted) Renamed ingestion => 3. Creating Lance Datasets (and also expanded it)

Added "hello world"

This requires the pandas ImageDtype extension so I won't merge this until that's done.

* make it runnable

* address PR comments

Co-authored-by: Lei Xu <lei@eto.ai>