Add lance.util.duckdb to help install the extension transparently (#301)
* Add lance.util.duckdb to help install the extension transparently

It's too complicated to make users go through the installation process
themselves. Instead let's create a small python utility function.
In a python-centric workflow, this will be the easiest method. Alternatives
can also be considered like providing a torch-like website or a command line
utility that the user needs to download.

* fix

* don't run mac tests twice

* address PR comments

* skip mac ext test since we can't get around virtualization issue

* check duckdb version and raise friendlier exception if mismatch