refactor: move main macro benchmarks to python (#1235)
Fixes https://github.com/lancedb/lance/issues/1218

This makes it easier to write, run, and profile benchmarks. Writing them
in Python makes them faster to compile and not require recompilation if
we only modify the benchmark itself.

We cache the expensive fixtures (like an indexed table) so that we can
quickly run the benchmarks.