fix: search should not panic on empty table (#1289)
This PR add a test for KNN search on an empty table. Our code previously
returns error because `KNNFlat` returned an empty record batch without
`_distance` column. This PR makes `KNNFlat` return a schema with
`_distance` column.

This PR also adds a test for KNN search on an empty table that was
previously non-empty. This also seems to fail, but is not fixed in this
PR. I created a new issue
[here](https://github.com/lancedb/lance/issues/1288)

This PR sneaks in `location: Location` field to `crate::Error::Schema`.
I needed this for debugging.