fix: knn search works after compaction (#1304)
Fixes #1303 

After compaction, the fragment ids may not be in sequential order. We
were using
`binary_search` to re-order results received from parallel
`Fragment::take_rows()` calls,
which requires sorted data. This PR switches to just doing linear search
for now, which
doesn't require sorting.

We should optimize this code path better in the future, but the batches
are small enough for
linear search is likely not too slow.