feat: add naive prefiltering (#1276)
This adds a prefilter option to the list of scan options. If prefilter
is true then the filter is applied before the index (instead of after
the index). Currently this only works when using a flat KNN scan (e.g.
when use_index is false) and so it may be very expensive on large
datasets if the filter is not very selective.

Closes #1032

---------

Co-authored-by: Weston Pace <pace@Westons-MacBook-Air.local>
Co-authored-by: Will Jones <willjones127@gmail.com>