perf: improve scan with filter performance (#1252)
This PR improves the performance of scanning with a filter. This uses
late materialization in Lance, which is an entirely different code path
from scanning without a filter. It first reads the filter columns,
applies the filter, and then calls `take_rows()` to get the rows that
matched the filter.

Two major optimizations have been made here:

1. `take_rows()` gets two new optimized code paths: one for when the
indices are contiguous and one where they are sorted.
2. `BinaryDecoder::take()` is refactored so all data goes through only 1
memcpy and not 2. Before, we called `take()` on all ranges and the
`concat()` to bring all the ranges together. This PR changes so we copy
binary data into a perfectly pre-sized buffer. When we detect a range is
contiguous and non-repeating, we just copy the full range instead of
iterating through the values and copying each sub-value.