docs: document commit API as basis for distributed writes (#1157)
This is a first pass at providing "distributed" examples. I ran out of
time to do anything more elaborate. For now, we demonstrate the
distributed APIs just in a REPL, which helps convey the concepts. Later,
we can add notebooks showing how these integrate into a framework like
PySpark or something else.

This PR makes `LanceDataset.commit()` a public API. We already have
users using it and we've just made the API improved, so I think this is
something we'll want to do. It is labelled as an advanced API in various
place though.

Closes #1179