chore: add pre-commit config file to make it easier for future devs to setup python commit hooks (#1319)
Pre-commit hooks for linting and formatting are useful but require
manual configuration. [pre-commit](https://pre-commit.com/) is a python
tool that automates the installation of pre-commit hooks from a config
file. This PR adds a pre-commit config file and instructions on how to
configure the pre-commit in the python developers guide.

These changes do nothing automatically. Users still need to manually
download and install the pre-commit tool.