fix: make PageTable handle multiple files (#1332)
Currently, `PageTable` assumes there are fields `0..num_columns`, where
`num_columns` is the total number of columns in the dataset. This works
for new files added, because we always write `0..(max_field_id + 1)`
entries in the page table, but it breaks for old data files who have
fewer fields. This is why after merge we can read the new columns (in
the new files), but not the old columns. Also, because the PageTable is
cached, we could read while using the same instance of `LanceDataset`,
but as soon as we loaded in a new session, the PageTables would be
re-read and break.

This PR modifies the `PageTable` format slightly so that the fields
start at the minimum field id in that file's schema. This is backwards
compatible for all tables that only have one data file per fragment, but
breaks tables that have more than one file. However, seeing as the
latter case is broken already, I think this breaking change is fine.

Closes #1331.