feat: add a cache for dynamodb schema validation (#1308)
Prior to this PR we describe dynamo db table everytime opening a
dataset. This is okay when the number of calls to opening a table with
dynamodb store is low. As `DescribeTable` has a account level limit of
2,500 RPS, we should avoid repeatedly calling the API.

This PR also fixes error formatting for aws sdk errors. AWS sdk error
doesn't include the underlaying error unless asked for.