chore: cleanup tests to avoid errors when optional components are not present (#1374)
There are several mildly related changes here:

* I don't have torch installed in my environment and these supposedly
optional tests were failing. Even though the tests were checking for
this that check was not working correctly because `import torch` was
importing the `python/tests/torch` directory as a module. So I renamed
this to `torch_tests`.
* Marks allow you to make tests "opt out" but, for these optional
features (like integration tests) I think it's more intuitive to have
them be "opt in". I changed the way the integration test mark is
configured to allow for this.
* There was a cuda mark but I removed it in favor of checking if `torch`
is installed. Although, I don't think these things are exactly the same.
It's possible that a machine has torch but not cuda (because it has some
other GPU type). Ideally we would be able to detect if cuda is present
on the machine in some way and programmatically disable those tests (in
the `pytest_collection_modifyitems` method) similar to how the
integration mark is configured.
* I changed `make test` to now include `-W error::FutureWarning` so that
we can detect things like https://github.com/lancedb/lancedb/issues/548
more quickly