feat: collect row ids and run index remapper (#1377)
During the commit_compaction step (the last part of compaction, after
the fragments have been rewritten) we run an index remapper to remap the
indices. This is needed because the row addresses will have changed as
part of the compaction. We should only remap indices that had rows in
affected fragments.

I also removed the "rewrite operation" from the python bindings. This
was already a fairly complicated operation and adding rewritten indices
to it would make it even more complicated. I don't think users are going
to be orchestrating compaction themselves anytime soon (and we can add
it back in if needed). Furthermore, the plan compaction capability and
serialization of compaction tasks is a much better solution for anyone
that wants to orchestrate the compaction process.

---------

Co-authored-by: Will Jones <willjones127@gmail.com>