perf: better parallelism in delete vector prefiltering (#1428)
The previous handling of delete files in an ANN search was sub optimal
in a few ways that are fixed by this PR:

* The synchronous portion of applying the prefilter was not run in
parallel across the different PQ pages. Now it is run in parallel as
part of the spawn_cpu critical section.
* The synchronous work of calculating the prefilter (converting from
delete file to, previously, a hashmap) was repeated for each PQ page.
Now we run this once, in parallel with the IVF centroid calculation and
any IVF/PQ page loading.
* The prefilter was using a computationally heavy HashSet. Now we use a
roaring treemap.