test: fix flaky tests involving tokio::fs::File (#1430)
When a tokio::fs::File is dropped it does not wait for any ongoing
writes to finish (as this would be a synchronous blocking wait).
Instead, it is recommended that you call `shutdown` or `flush`. This was
causing some unit tests to fail intermittently because we were creating
a `File`, writing to it, then immediately reading from it. Sometimes
these reads would fail because the data had not actually been written
yet.