perf: use selection vector strategy to improve exact knn performance with deletions (#1418)
This PR repurposes the `_rowid` column's validity buffer as a selection
vector. By using a selection vector, we can defer the mem copies
involved with applying deletions to columnar data, which is a big
bottleneck in KNN queries with deletions.

For convenience, we make the `_distance` column nullable. The final
output will never be null, but we sometimes will have the intermediate
values be null.

Closes #1352

TODO: 
* [x] switch to using `_rowid` validity bitmap as selection vector.