fix: fix an issue where the GPU index trainer was taking too much data into memory (#1447)
the old computation calculated `k * sample_rate` as `sample_size` and
then took the whole dataset unless there were `k * sample_size` rows
which is `k * k * sample_rate`. I've reworded the logic to hopefully be
a bit clearer.

Once this was fixed the next issue was that the `dataset.sample` method
doesn't actually exist and so I added it.