feat: allow prefiltering to be used with an index (#1435)
Currently, prefiltering is restricted to searches that do not use an
index. This PR enables prefiltering to be used on an indexed column. The
approach is to first run a query to figure out the valid row ids and
then pass this list of valid row ids to the index.