feat: add parameter to split by file size (#1444)
It's possible to hit the max size for object-store (100GB). (This limit
is set by the multi-part upload implementation, not the underlying
stores themselves.)

To help users avoid this, we provide a new `max_bytes_per_file` settings
in `WriteParams`. This is checked as we flush each batch.

For users of the `Fragment.create()` API, we provide an alternative
called `write_fragments()` that can split into multiple fragments if
relevant limits are hit.