feat: add btree scalar index (#1476)
This PR fleshes out the concept of a scalar index. It creates a
btree-inspired index as an initial example index. The index is modeled
very similarly to the vector index and will also closely resemble an IVF
index.

The index is broken into two parts, the lookup table / btree and the
data pages (leaf nodes). The lookup table is stored on disk as a lance
table with a single record batch containing the min/max/null
count/offset of each data page (the offset is currently superfluous and
I'll probably remove it as I clean this up). The data pages are stored
on disk as a lance table with one record batch per page. When we load
the lookup table into memory we convert it into a btree.

An index search first uses the btree to figure out which pages might
satisfy the query. We then load those pages into memory and search
through them with a flat scan. In the future we could investigate
storing leaf pages in some b-tree like structure but since they are
likely to be small enough to fit into a single IOPS I'm not sure there
would be much benefit.

To train, the btree index needs a stream of batches. These batches
should be rechunked into smallish chunks. Each batch should contain two
columns, the value and the row id. The batches should be sorted by the
value column. There is very little compute and the entire operation will
probably be I/O bound and fairly quick (however long it takes to sort
and scan the data) but memory intensive (no out-of-core sort yet).

The one parameter we have is the size of these data pages. For example,
we could decide each page is 4096 rows. This would mean, with 1Bi rows,
we would have 256Ki pages. Our lookup table should still be fairly small
(e.g. if the values are 4 byte integers then the lookup table will be
~3MiB.)