feat: make it possible to create and load scalar indices for a dataset (#1516)
Three new internal methods are added to the dataset (via `DatasetIndexInternalExt`):

 * `open_generic_index`
 * `open_scalar_index`
 * `open_vector_index`

Previously, there had only been `open_index` which was a free function
(that took a `&Dataset`) and returned a `Arc<dyn VectorIndex>`. That
method still exists as an internal helper but users should now use
`Dataset::open_vector_index` instead.

In addition, there is a new `IndexType` for `Scalar` and the
`create_index` method will recognize this type. Creating a scalar index
currently does not require any parameters but it may in the future.

Opening a vector index today is rather involved. First the proto
metadata is loaded from disk. This describes the detailed
vector-specific parameters of the index (the index information in the
manifest is very high-level and common across vector and scalar index
types). The scalar index does not need that today because we can assume
all scalar indices are btree backed by flat. As a result, we did not
need to make any changes to any .proto files. In the future, should we
add new types of scalar indices and new parameters, we should be able to
handle that similarly. These could either be added to the existing
`index.proto` or maintained in their own `scalar_index.proto` (or even
stored in some other format entirely).

The generic `Index` trait has moved into `lance-index` and is required
by both `ScalarIndex` and `VectorIndex`. This trait has all of the
"common" index operations. Right now this is only a single method to
fetch the statistics but in the future it will hopefully include things
like remap and update/optimize.

The index statistics were previously returned as a `serde_json::Value`.
This is now returned as a simple `String`.