fix: fix a memory leak where a dataset would not be fully deleted (#1606)
If you create/open a dataset, search the dataset (using a vector index)
and then drop the dataset it will leak some memory. This is because the
dataset has an index cache. The index cache keeps strong pointers to the
indices. However, in the IVF case, we were also keeping a strong pointer
to the index cache. This caused a cycle and the index and cache kept
themselves alive.

This fix downgrades the IVF index's pointer from `Arc` to `Weak` to
break the cycle. I've verified the fix manually but am not sure of a
good way to detect this sort of thing in testing. Perhaps we can
investigate running valgrind?