chore: minor changes to tracing instrumentation (#1619)
These are pretty minor fixes but I am trying to make sure that all the
spans we have are debug level except for those at the dataset/scanner
layer (which are info). Then, I am trying to make sure that all of our
debug scans always roll up to one of the info spans. This generates
better traces as there are fewer (and more meaningful) root spans.