perf: use datafusion to shuffle index partition data (#1645)
This removes the current `Shuffler` in favor of using DataFusion to sort
the index data into partitions. DataFusion supports spilling to disk
while sorting larger-than-memory data, but will keep the data in memory
if it can fit. In theory this can achieve good performance when we need
to shuffle the data.

To keep this to a small refactor, I've only replaced the shuffler.
However, we could expand the role of DataFusion further by having the
partition ID assignment and PQ transformations written in terms of
DataFusion's DataFrame API.

Closes #1483