fix: raise the manylinux version to 2_24 on aarch64 linux builds to avoid cross compilation issues (#1737)
The manylinux 2_17 image does not seem to be generating a working build
of the `ring` library. We had originally managed to get it to compile by
setting `export CFLAGS_aarch64_unknown_linux_gnu="-D__ARM_ARCH=8"`
however the library that was produced didn't work quite right and would
generate "bad signature" when working with S3. By bumping to manylinux
2_24 we are able to generate a successful build. If someone needs 2_17
then they will need to dig into the `ring` library failure in more depth
to figure out the exact cause of the issue.