feat: allow externally generated shuffle buffer and multiple of them (#1801)
This PR does a few things to the IVF shuffler

* add a `set_unsorted_buffers` method. This method allowes caller to set
a list of buffers that is different from the `unsorted.lance` buffer
used by default.
* This is helpful for using externally calculated buffers (say buffers
from multiple GPUs)
* This also removed the limit of the number of rows that can be
shuffled. Previously, we could only shuffle as many as 2^32-1 rows as we
used a single shuffle buffer with a limit of 2^32 rows
* added logic to handle multiple shuffle buffers, by default we only use
one buffer
* added a bunch of tests for the shuffler.

I will wire up python in the next PR to keep this smaller and more
reviewable.