fix: spurious bug during compaction (#1829)
During compaction we read in the row ids so we can use them for
remapping. We read the row ids into a tree map which is most efficient
if we can read them in order and the current code is in fact assuming
that the row ids arrive in order. However, we weren't actually scanning
in order and this could lead to failures during compaction.