refactor: split any reference to fragment or manifest out of the file reader (#1848)
I'm working on splitting the file format and the table format into their
own crates. One point of difficulty was that the file reader referred to
the fragment / manifest for loading deletion vectors and other
information. This PR brings all of that up into the fragment level.

The deletions are still passed to the file reader as a deletion vector,
so this PR doesn't change where the work happens. However the argument
is now provided at read time instead of at open time.

There are no API changes and should be no changes to the format or how
things are interpreted. This is purely a refactor.