refactor: split lance-core into lance-core, lance-io, lance-file, and lance-table (#1852)
There is very little code change in this PR. It is primarily moving
files around.

The only substantial change should be a change to the `FileWriter`. It
is now either `FileWriter<NotSelfDescribing>` (for testing in
`lance-file`) and `FileWriter<ManifestDescribing>` for use everywhere
else. This is because a lance file writes the schema as a manifest and
manifest is part of the table format. In the future we could fix this by
creating a schema message in the protobuf in lance-file and changing the
`FileWriter` and `FileReader` to read this. However, this would be a
breaking change and would require a bump in the lance file version (or
feature flags). I'd prefer to save that change for later.