fix: merge insert operation fails if any of the payload columns are non-nullable (#1899)
This is because we do an outer join and the output columns of an outer
join are always nullable. As a result we try and insert nullable data
and get a schema mismatch. This fix avoids this problem by restoring the
schema after we select the appropriate rows from the outer join output.