fix: avoid a buffer copy and share the array to reduce RAM cost (#1907)
The centroids array can sometimes be quite large (on my test case it was
1536 dimension * 32Ki parts * 4 B = 192MiB. When we compute partitions
we do so in num_cpus^2 tasks (on my system this is 16 * 16 = 256).

We were copying the centroids data to each task (so for my test this
required 48GiB of RAM). Since this array is read-only, we can just make
a shallow copy of the arrow array instead of copying the data.