fix: only read the data file's fields from the page table and not the whole dataset's fields (#2095)
During the refactor to split things into lance-file and lance-table I
introduced a bug. There were a number of code paths for opening a file
reader and I consolidated them into one. However, when we were reading
one data file from several data files, we ended up trying to load too
many fields from the page table (more fields than exist in the page
table).

Most of the time, this was ok. The read would read past the end of the
page table and grab part of the schema (effectively garbage data). We
never actually accessed this data and so we would get away with it.
However, if the fragment schema was considerably larger than the data
file schema, then the read would go past the end of the file and cause
an error.