fix: use the fair spill pool instead of the greedy spill pool (#2126)
Right now we use the greedy spill pool when we use spilling for btree
index training. This is because the btree index plan is:

```
Input Stream -> Sort    -> SortPreservingMerge -> Output
Input Stream (old data) -/
```

My thought was that `Sort` was the only spillable operator and we could
use the greedy pool. It turns out that `SortPreservingMerge` is also
spillable and we can't get away with the greedy pool.