feat: make substrait optional to allow avoiding libgit2 (#2168)
Hide substrait behind a feature. This will remove the libgit2 dependency
in the short term. In the medium term this won't be needed because
datafusion's next release should include substrait 0.29.3 which removes
libgit2. However, we still might want to keep substrait behind a feature
just to stick to a general spirit of keeping things lean.