feat: add v2 nullability for primitive / fsl (#2169)
Struct and list will come later. This PR was getting large enough as it
was.

There is some refactoring that is happening as part of this PR also. The
basic encoder was previously and array encoder that took in two buffer
encodings. However, this meant we couldn't use the basic encoder to
encode FSL since FSL is an array encoder. So I changed the basic encoder
to be an encoder that takes an encoder for validity and an array encoder
for values.

I then also created the "value encoder" as an array encoder and renamed
the existing value encoder (a buffer encoder) to the flat encoder.

In addition there was some minor refactoring needed to keep track of
buffer indices. Previously array encodings always mapped to 2 buffers.
Now they might map to more than 2 buffers (e.g. FSL with its own
nullability is 3 buffers).