refactor: move existing utilities from file reader into fragment-level operations (#2197)
Applying row ids and deletion vectors is currently a file-level
operation. However, these are table-level concepts and I need to move
them to the fragment reader (the v2 file reader doesn't use these
concepts).

Eventually, we may like to pushdown deletions (to get "free"
materialization), but that can wait until we are ready (I'm not sure we
will reuse DeletionVector and may use something simpler? Also, it'll be
a while before we care about this particular optimization I think)

Note: this PR does not remove the source utilities (e.g. we are not
changing the file reader). That will come in a future PR.