refactor: migrate scan path from v1-only to generic reader (#2201)
This fleshes out a "generic reader" trait added in a previous PR and
implements the trait for the v1 reader. It then adapts the scan path to
use the generic reader. There are still many paths (take, update, etc.)
which use the legacy read path directly. These can be migrated over more
slowly as we go.

Next, I will add an implementation for the v2 reader and this should
make it possible to use the v2 reader in the scan path.