feat: replace `LanceFragment.add_columns` with `merge_columns` (#2208)
`LanceFragment.add_columns()` is broken for schemas with complex row id
sequences. It needs to return the actual Lance schema used, rather than
allowing the user to pass a PyArrow schema. PyArrow schemas are
problematic because they assume field ids are contiguous and sequential,
but they often are not. So we introduce a method
`LanceFragment.merge_columns()` which returns both the new fragment and
the actual Lance schema.

* Added new `lance.schema.LanceSchema` class, representing a Lance
schema.
* Added `LanceDataset.lance_schema` property that returns a
`LanceSchema`.
* Deprecated `LanceFragment.add_columns()` in favor of
`LanceFragment.merge_columns()`, which returns `Tuple[FragmentMetadata,
LanceSchema]`.
* Also motivating this change: I'd like to later re-introduce
`LanceFragment.add_columns()` with a signature in line with the
`LanceDataset.add_columns()`.
* `LanceOperation.Merge()` now accepts `LanceSchema` in the `schema`
argument. `pa.Schema` is still accepted for now but issues a
`DeprecationWarning`.
* `lance.debug.format_fragment()` now requires `LanceDataset` as its
second argument.
  * This is a brand-new debug API, so I think it's fine if we change it.
* Changed some fragment methods to use
`FragmentMetadata.from_metadata()` instead of
`FragmentMetadata(frag.json())` to avoid silly JSON roundtrip.
* `LanceDataset.commit()` now runs basic validation on fragments to
prevent some corruption. I will add more checks like this in the near
future.
* `FragmentMetadata` no longer holds a schema. pickle/JSON serialization
didn't preserve it, so support for it was already broken. It didn't seem
to be relied upon by anything.