feat: support complex schemas in append (#2209)
Fixes two bugs, both associated with schemas that have holes in the
field ids:

1. `write_fragments()` and `LanceFragment.create()` assume they can
derive the field ids from the Arrow schema. This is not the case if
there are holes in the schema. Therefore, when the mode is `Append`, we
check the existing schema of the dataset and use its field ids. Fixes
#2179
2. `PageTable` assumed that there were no holes in the field ids. It was
parametrized as `field_id_offset` and `num_fields`, assuming the field
ids were `field_id_offset..(field_id_offset + num_fields)`. This is
changed to be parametrized by the min and max field id.

---------

Co-authored-by: Weston Pace <weston.pace@gmail.com>