chore: remove --preview from ruff invocations (#2220)
We had originally added --preview to our ruff invocations because it was
required to enforce rule CPY001 which checks for a copyright header. We
have since moved that check to a dedicated tool and removed the ruff
rule so we can safely remove --preview.

The --preview flag was a problem because IDEs and manual invocation
would often leave it off and it turns out that the formatting changes
based on the presence of `--preview` (it doesn't just enable / disable
rules but actually changes the behavior of rules too).

By removing --preview we should make it easier for developers because
there will be less configuration of their development environment
required.