fix: copy arrays when placing them in the v2 writer's accumulation queue (#2249)
When a record batch crosses the FFI layer we lose the ability to
deallocate that batch one array at a time (every array in the batch
keeps a reference counted pointer to the batch). This is a problem for
the writer because we usually want to flush some arrays to disk and keep
other arrays in memory for a while longer. However, we do want to
release the arrays we flush to disk to avoid memory leaks.

This issue was highlighted in a test / demo that attempted to write
video data to a lance file. In this situation we were writing the data
one row at a time. There were only 30,000 rows in the source data but
this was hundreds of GB. None of the metadata columns would ever flush
to disk (e.g. a 4 byte int column @ 30,000 rows is only 120KB). These
metadata arrays were keeping the video data alive in memory and the
writer was taking up too much data.

The solution in this PR is to perform a deep copy of any data we are
planning on flushing to disk. This is unfortunate, and there is a
configuration parameter to disable this copy (I have, intentionally,
chosen not to expose this in the python API since data from python
always crosses an FFI boundary and is susceptible to this problem).
However, copying this data by default is a much safer course of action.

In the future we could investigate alternative ways of passing the data
across the FFI boundary (e.g. instead of sending the entire record batch
we could send individual arrays across and then reassemble them into a
record batch on the other side). However, I don't think we need to worry
too much about this until we see writer CPU performance become an issue.
In most cases the batches will probably be in the CPU cache already and
so this should be a pretty quick write. Also, writers that are writing
any significant amount of data will be I/O bound.