feat: raise GCS object size limit to 2.5TB (#2261)
This is an interim fix for just GCS to solve
https://github.com/lancedb/lance/issues/2247

Because of the challenges of casting between `&dyn ObjectStore` and
`&dyn MultiPartStore`, it's not easy in the current version of
`object_store` to implement this generically over all stores. However,
in the next version of `object_store` (0.10.0), there is a new API for
`put_multipart()` that will make it easy to extend this implementation
to all stores.