fix: don't use pushdown scan if a custom batch size has been set (#2277)
The pushdown scan currently requires fully materializing a batch in
memory. As a result, it is not capable of supporting the `batch_size`
parameter in the scanner. This fix is not great, it disables pushdown
optimization when a batch size is specified, but if users truly need a
batch size (to constrain memory) then its better to give them what they
ask for at higher cost.

I would like to do a more correct fix soon but probably not until I work
on pushdown for v2 (I'm still not sure if the "late materialization"
will be part of the file reader itself or if we still need a dedicated
pushdown node.