chore: simplify the datagen API (#2288)
Very minor change but one that has been annoying me for a while. The
datagen allows columns to be specified with or without a name ("without
a name" means one is automatically generated based on the column index).
If a name is specified it must be a `String`. The vast majority of usage
does specify a name and provides it as a `&str`. As a result we have a
lot of boilerplate `Some("foo".to_string())` that I would like to avoid.

To generate a column without a name use the function `anon_col`.
When generating a column with a name the name is now provided as `impl
Into<String>` instead of `Option<String>`.