feat: add projection support to the v2 format (#2296)
We can simplify the projection a bit more down the line. With this PR
you must specify both:

 * The column indices to load
 * What data types you want to load into

This is because the file format doesn't technically have a type system.
Each column is just a bunch of data encoded in various ways. So, for
example, you could interpret a column as `List<u8>`, or `String`, or
`LargeString`, or `Binary` or `LargeBinary` or `StringView` or
`BinaryView` or ...

However, at some point we may want to create a "pick a default column to
decode into based on the type" extension point that can be extended by
users (this needs to be extensible since encodings are extensible) which
picks an appropriate type based on the encoding. This is slightly more
complicated than it seems because there can be multiple encodings for a
single column and because encodings may be user-defined. This way users
will only need to provide the column indices to load.

In the short/medium term this is not too urgent. The Lance table format
has a type system and stores the schema and so it can easily figure out
both pieces of required information.