feat: add priority to I/O scheduler (#2315)
This also renames store scheduler to scan scheduler. I'm thinking I
don't want to get into the thorny issue of how to prioritize I/O
requests across different scans. So, with this approach, if multiple
scans are running at the same time, then they will overschedule (and let
the OS deal with it). This can be revisited in the future.

Closes #1958