ci: fix labeler permissions (#2348)
Earlier in
https://github.com/lancedb/lance/pull/2332#issuecomment-2110173164 we
switched to `pull_request` events to checkout the correct PR. However,
in those actions, GHA doesn't have permissions to label PRs, so that
change broke the labeler job. This switches back of
`pull_request_target`, but adds code to checkout the correct branch when
checking for breaking changes.