fix: rework decoding to fix bugs in nested struct decoding (#2337)
This PR modifies both scheduling and the decode stream. These
modifications should mainly affect the logical encodings (list /
struct). The changes to the other encodings are fairly minimal.

In the scheduler we now track where we are in the field tree. This is
both for debugging purposes (so we can log in trace messages the current
path) and because we now need to send the path as part of the message we
send to the decode stream.

The decoder changes are more significant. Previously, we combined
waiting for I/O and waiting for additional encoders into the same phase.
This logic was more complex, but more importantly, it also assumed that
the decoder could recreate the order in which the scheduler scheduled
pages. For example, if we scan through the columns and encounter one
that needs more data, then we grab exactly one page, and continue the
column scan, grabbing the next page when we come by in another pass.
This works in many cases but doesn't work in others. For example, the
scheduler might schedule the many pages in a row for the same column if
that column is wide or the page size is small. The decoder would get out
of sync in these cases.

Now, the logic is simpler, and more correct. The decoder first waits for
enough scheduling to be done that a batch can be delivered. During this
pass we drain encoders from the scheduler and insert them into the
current decoders not by "current position in the decode" but by the path
that is now included with the decoder.

Once enough scheduling has been done we then wait for I/O.

Once I/O is done we then drain the decoders in the same fashion we did
before.