fix: rework list encoder to handle list-struct (#2344)
We had previously scanned through the items pages triggering decoding in
a manner that was very similar to the struct decoder but incomplete.

Now, we reuse the struct decoder in the list decoder. So the "indirect
schedule & decode" is nearly identical to a direct schedule & decode.
This removed some near-duplicate code and makes it so that we can handle
list-struct (or list-anything)

This also forced us to fix another issue that was preventing us from
decoding files with more than 2^32 rows.