feat: general compression for value page buffer (#2368)
This PR introduces general compression for value page buffers, starting
with zstd, to reduce the on-disk size of all types of value arrays.

Here are the key details:
1. After some code exploration, I implemented this as a buffer encoder
for `ValueEncoder` instead of as an independent physical encoder. Please
let me know if this approach is suitable.
2. Enhancements to `ValuePageScheduler.schedule_ranges` allow it to read
the entire buffer range for compressed buffers. To support this, I added
a new `buffer_size` metadata to several buffer structs, populating these
variables using metadata from Lance.
3. Currently, only zstd compression with the default level (level 3) is
implemented. If this approach is deemed suitable, we can consider adding
more general compression methods in the future.
4. In a specific test case, the original Lance file was 13MB. After
applying zstd compression, its size was reduced to 7.4MB.

---------

Co-authored-by: Weston Pace <weston.pace@gmail.com>