feat: make the index cache size (in bytes) available (#2381)
This PR makes extensive use of the deepsize crate for tracking the deep
size of standard structures (Vec, BTreeMap, etc.). This crate does
appear to be abandoned (no updates in 3 years). However, I'm unable to
find anything similar, the code is fairly small should we need to vendor
it in the future, and it is unlikely to be doing anything dangerous.

It turns out we store many many things in the cache (transitively)
though most of them are probably small.