feat: add special uri scheme to use CloudFileReader for local fs (#2402)
This PR adds a new uri scheme, `file-object-store` to lance. When
reading a file, we branch and use `LocalFileReader` if the object store
is local. This is to reduce overhead and improvement performance of
local datasets.

However, this makes local file system harder to test as it uses
different code path for read. This new scheme will force local dataset
use `CloudFileReader` so we can test behaviors unique to remote storage
on local fs.

Note: doesn't work on window. but `file://` doesn't work on windows
anyways