refactor: shuffle around v2 metadata sections to allow read-on-demand statistics (#2400)
Previously the v2 file format had described the column metadatas and
column metadata buffers in the same part. This is problematic because
there is no way to know when the proto stops and the metadata begins.
It's also makes it more difficult to do "read on demand" column
statistics (and dictionaries). This PR gets rid of column metadata
buffers. If a column encoding wants to refer to a data buffer of some
kind that data buffer can be written to the data section. This helps
account for things like dictionaries which might need to be periodically
flushed to the disk. We also move the global buffers here for
consistency.

Now a reader can quickly read all column metadatas and the schema in a
single IOP and then follow that up by loading column-specific buffers on
an as-needed basis. This will allow us to only load statistics /
dictionaries for columns that we are interested in.

---------

Co-authored-by: Bert <albert.lockett@gmail.com>